requirements.txt
tensorflow-cpu
tensorflow-addons
tf2onnx
